tools/requirements.txt
pyexiftool==0.5.6
